Add check/test/clippy/fmt pipeline

+++ src/dct.rs
@@ -1,8 +1,7 @@
 use std::f32::consts::{FRAC_1_SQRT_2, PI};
-use crate::log;
 
 pub fn spatial_to_freq(block: &[[u8; 8]; 8]) -> [[u8; 8]; 8] {
-    let mut result = [[0 as u8; 8]; 8];
+    let mut result = [[0_u8; 8]; 8];
 
     for x in 0..8 {
         for y in 0..8 {
@@ -13,6 +12,7 @@ pub fn spatial_to_freq(block: &[[u8; 8]; 8]) -> [[u8; 8]; 8] {
     result
 }
 
+#[allow(non_snake_case)]
 fn G(u: usize, v: usize, block: &[[u8; 8]; 8]) -> f32 {
     let (u, v) = (u as f32, v as f32);
     let au = if u == 0.0 { FRAC_1_SQRT_2 } else { 1.0 };
